Change buffer wrappers to use structs
Idea is to cut down on the usage of delegates

+++ src/Shared.Desktop/Registry/Types/RegistryKeyHandle.cs
@@ -37,7 +37,9 @@ namespace WInterop.Desktop.Registry.Types
 
         private RegistryKeyHandle(uint predefined, bool isPerfKey = false) : this(ownsHandle: false)
         {
-            handle = (IntPtr)predefined;
+            // Need to cast to int first to avoid Overflow when
+            // casting to IntPtr when running on 32 bit.
+            handle = (IntPtr)(int)predefined;
             IsPerfKey = isPerfKey;
         }
